Restructure course and language export
